refactor: move anm operator in anm packed
